requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.4
Django==4.0.6
djangorestframework==3.13.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
